Removed a compiler warning.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@664 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/particles.c
@@ -408,7 +408,7 @@ else {
 			pParticle->nHeight = (short) ((double) pParticle->nHeight * decay);
 			}
 #endif
-		pParticle->nLife -= t / gameStates.gameplay.slowmo [0].fSpeed;
+		pParticle->nLife -= (int) (t / gameStates.gameplay.slowmo [0].fSpeed);
 		}
 	}	
 pParticle->nMoved = nCurTime;
